+++ test/cases/require-styles-inline-resolve-url/webpack.config.js
@@ -1,0 +1,56 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  resolve: {
+    alias: {
+      Views: path.join(__dirname, 'src/views/'),
+      Images: path.join(__dirname, 'src/assets/images/'),
+      Styles: path.join(__dirname, 'src/assets/styles/'),
+    },
+  },
+
+  entry: {
+    index: './src/views/index.pug',
+    'pages/about': './src/views/about.pug', // test same inline CSS from different paths
+  },
+
+  plugins: [
+    new PugPlugin({
+      extractCss: {
+        filename: 'assets/css/[name].[contenthash:8].css',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+
+      {
+        test: /\.(png|jpg)/,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+};